Switch to Fedora 36 and sync files over NFS
The current directory is mounted on `/vagrant` by default
and it's only `rsync`ed on startup. We can run `vagrant rsync-auto`
on another shell window but it's a bit too cumbersome for us.

Given we're not after maximum performance, we can make do with NFS
with no problems at all!

We should look into disabling the initial `rsync` synchronizaiton
though!