Add VLAN ID to Veth names
We need to somehow include the VLAN ID into a Veth's name.
Otherwise, we're bound to try and reuse a Veth name when a
host belongs to more than a single VLAN. This scenario would
prevent the network from being instantiated altogether...

This change is compatible with previously desgined networks!